transcoder/requirements.txt
fastapi
uvicorn
aiosqlite
apscheduler
python-dotenv
